orders

+++ app/Http/Controllers/Admin/GroupController.php
@@ -8,10 +8,10 @@ use App\Http\Requests\Group\GroupUpdateRequest;
 use App\Models\Category;
 use App\Models\Group;
 use App\Models\Product;
+use Illuminate\Http\Request;
 
 class GroupController extends Controller
 {
-
     /**
      * Display a listing of the resource.
      *
@@ -20,9 +20,9 @@ class GroupController extends Controller
     public function index()
     {
         if (auth()->user()->role == 'admin') {
-            $groups = Group::with(['category:id,title_rus', 'products:id,group_id,preview_image'])->paginate(5);
+            $groups = Group::with(['category:id,title_rus', 'products:id,group_id,preview_image'])->paginate(8);
         } else {
-            $groups = auth()->user()->groups()->with(['category:id,title_rus', 'products:id,group_id,preview_image'])->paginate(5);
+            $groups = auth()->user()->groups()->with(['category:id,title_rus', 'products:id,group_id,preview_image'])->paginate(8);
         }
         return view('admin.group.index_group', compact('groups'));
     }
@@ -34,7 +34,11 @@ class GroupController extends Controller
      */
     public function create()
     {
-        $products = auth()->user()->products()->get();
+        if (auth()->user()->role == 'admin') {
+            $products = Product::select(['id', 'title'])->get();
+        } else {
+            $products = auth()->user()->products()->select('id', 'title')->get();
+        }
         $categories = Category::all();
 
         return view('admin.group.create_group', compact('products', 'categories'));   
@@ -65,11 +69,12 @@ class GroupController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  \App\Models\Group  $group
      * @return \Illuminate\Http\Response
      */
     public function show(Group $group)
     {
+        $this->authorize('view', $group);
         $group->load(['products:id,group_id,title,preview_image', 'category:id,title_rus']);
         return view('admin.group.show_group', compact('group'));   
     }
@@ -77,14 +82,19 @@ class GroupController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  \App\Models\Group  $group
      * @return \Illuminate\Http\Response
      */
     public function edit(Group $group)
     {
+        $this->authorize('update', $group);
+        if (auth()->user()->role == 'admin') {
+            $products = Product::select(['id', 'title', 'group_id'])->get();
+        } else {
+            $products = auth()->user()->products()->select(['id', 'title', 'group_id'])->get();
+        }
         $group->load(['category:id,title,title_rus']);
         $categories = Category::all();
-        $products = auth()->user()->products()->select(['id', 'title', 'group_id'])->get();
         return view('admin.group.edit_group', compact('group', 'categories', 'products'));   
     }
 
@@ -92,13 +102,14 @@ class GroupController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \App\Models\Group  $group
      * @return \Illuminate\Http\Response
      */
     public function update(GroupUpdateRequest $request, Group $group)
     {
+        $this->authorize('update', $group);
         $data = $request->validated();
-        $data['saler_id'] = auth()->id(0);
+        $data['saler_id'] = auth()->id();
         if(empty($data['products'])){
             $group->update($data);
         } else {
@@ -115,11 +126,12 @@ class GroupController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  \App\Models\Group  $group
      * @return \Illuminate\Http\Response
      */
     public function destroy(Group $group)
     {
+        $this->authorize('delete', $group);
         $group->products()->update(['group_id'=>null]);
         $group->delete();
         return $this->index();
